sample size calculation for different group sizes (Cohen 1988) and moved validation methods to separate file
